+++ CHANGELOG.md
@@ -3,6 +3,7 @@
  - Improvements on gemspec file and in ruby code using Rubocop (Thanks noraj)
  - Fixed issues with download and upload commands
  - Added support for simple wildcard usage in upload and download commands
+ - Fixed non-controlled error on wrong certificates path when SSL enabled
 
 ### 3.4
  - Added CI files to be used in Github actions for Docker automated building